requirements.txt
morphik==0.2.10
python-dotenv==1.1.1
pydantic==2.11.7
